Clarify ip6tables steps
Instead of providing the raw ip6tables.rules file, provide a .sh script (taken from http://www.dickson.me.uk/2013/03/15/setting-up-a-raspberry-pi-as-an-ipv6-gateway-using-hurricane-electric/) which generates the rules file. This way the user can more easily edit the rules.
By default allow SSH to the /64 and to the gateway.
Update readme to reflect the new steps